Correct Appending of Log Handlers
The current way logging is setup is that it appends a handler logging.getHandler('') on each __init__ of DashboardAPI() (and async classes).

I replicated the way logging is set in the BasicConfig currently to be set to the logger instance at the module level. This will allow it to function cleaner when included in applications that have already configured the root logger. The log levels are interesting. The logger level is set to DEBUG and pulled down to INFO for console logging when a file is also being logged to. If no file is specified, the log level is kept at DEBUG. This holds true in this patch so it can be reviewed in future changes. Notably, because the level is set to DEBUG in logger and not in the root logger, the log file (or console with no file) will not show DEBUG messages from other code (an example would be urllib3).