Logging inheritance, config file updates, and generator upgrades
Logging inheritance now supported thanks to @GameMaster47 and @ricardo.marques. Fixed some typos and aligned it with some PEP standards.

Updated use_iterator_for_get_pages kwarg to use config.py for the default False value, instead of a static False, to match the others kwarg. Thank you to @coreGreenberet for adding this option. The default remains False for backwards compatibility, but now that it's in config.py you can optionally override the default in your own instance of the library.

Upgraded generator to support the new 'sensor' type (Meraki MT product line). As of publishing this is available to beta orgs only. The MT sensors are pretty cool. You should try some.