Added path parameter escaping
This change filters URL path parameters through urllib.parse.quote to
sanitize inputs that are formatted into the url paths. This mitigates
the risk of path traversal attacks.

This also adds `administered` to the supported scopes for the newly added
/administered/identities/me (`getAdministeredIdentitiesMe`) operation.

This change also fixes a mix of tabs and spaces in the
`async_function_template` template.